#330 fix: ダークモードの表示を、スマホの場合はサイドバーに移動する
